# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stores/{id}/slots": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Post a new store slot",
                "description": "Create a new store slot",
                "operationId": "addSlot",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Store title",
                        "required": true
                    },
                    {
                        "name": "storeLine1",
                        "in": "query",
                        "description": "Store line 1",
                        "required": true
                    },
                    {
                        "name": "storeLine2",
                        "in": "query",
                        "description": "Store line 2"
                    },
                    {
                        "name": "zipCode",
                        "in": "query",
                        "description": "Zip code",
                        "required": true
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City",
                        "required": true
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Store country",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/slots/{slotId}": {
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store slot",
                "description": "Soft delete a store slot",
                "operationId": "removeSlot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    },
                    {
                        "name": "slotId",
                        "in": "path",
                        "description": "Slot id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/translations": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Post a new store translation",
                "description": "Create a new store translation",
                "operationId": "addTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store id",
                        "required": true
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/translations/{translationId}": {
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store translation",
                "description": "Soft delete a store translation",
                "operationId": "removeTranslation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    },
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}